Distribute keypoint geometry into existing backends; drop paz.keypoints
The JAX tree already organizes point geometry by math domain (points2D,
pointcloud, pinhole, poses); a dedicated keypoints backend cut across all of
them and duplicated points2D.transform, points2D.denormalize and
poses.project_to_image. Dissolve it into the backends that already fit.

- paz.poses: Pose6D, PnP solver constants, solve_PnP, solve_PnP_RANSAC,
  rotation_vector_to_matrix, solve_pose_matrix_RANSAC, project_points3D
  (next to project_to_image and the existing pose helpers).
- paz.points2D: normalize/denormalize_keypoints2D, rotate_point2D,
  rotate_keypoints2D, flip_keypoints_left_right, translate_keypoints, uv_to_vu.
  transform_keypoint is dropped in favor of the existing points2D.transform.
- paz.pinhole: build_cube_points3D (beside build_cube_corners).
- Callers (pose_estimators, pix2pose example, probabilistic keypoints example)
  import from these backends; paz.keypoints and its module are removed.

Tests move alongside: PnP recovery and project consistency in poses_test,
2D transforms in points2D_test, cube corners in pinhole_test.
