+++ examples/probabilistic_keypoint_estimation/train.py
@@ -20,7 +20,7 @@ def rotate_image_and_keypoints(key, image, keypoints, rotation_range):
     height, width = image.shape[0], image.shape[1]
     center = jp.array([(width - 1) / 2.0, (height - 1) / 2.0])
     image = paz.image.rotate(image, angle)
-    return image, paz.keypoints.rotate_keypoints2D(keypoints, angle, center)
+    return image, paz.points2D.rotate_keypoints2D(keypoints, angle, center)
 
 
 def translate_image_and_keypoints(key, image, keypoints, delta_scale):
@@ -28,7 +28,7 @@ def translate_image_and_keypoints(key, image, keypoints, delta_scale):
     scale = jp.array([width * delta_scale[0], height * delta_scale[1]])
     shift = jax.random.uniform(key, (2,), minval=-scale, maxval=scale)
     image = paz.image.translate(image, shift)
-    return image, paz.keypoints.translate_keypoints(keypoints, shift)
+    return image, paz.points2D.translate_keypoints(keypoints, shift)
 
 
 def augment_image_and_keypoints(key, image, keypoints, rotation_range,
